refactor: CommentItem / CommentThread 컴포넌트 분리
